requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
boto3==1.14.35
botocore==1.17.35
certifi==2020.6.20
cffi==1.14.1
cfgv==3.2.0
chardet==3.0.4
coverage==5.2.1
cryptography==3.0
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-appconf==1.0.4
django-awesome-avatar==1.1.3
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-cryptography==1.0
django-filter==2.3.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-oauth-toolkit==1.3.2
django-storages==1.9.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
docutils==0.15.2
filelock==3.0.12
gunicorn==20.0.4
honeybadger==0.3.0
httpie==2.2.0
identify==1.4.25
idna==2.10
isort==5.3.0
jmespath==0.10.0
lazy-object-proxy==1.5.1
Markdown==3.2.2
markdown2==2.3.9
mccabe==0.6.1
model-bakery==1.1.1
nodeenv==1.4.0
oauthlib==3.1.0
Pillow==7.2.0
pre-commit==2.6.0
psutil==5.7.2
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.5.3
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
virtualenv==20.0.30
whitenoise==5.2.0
wrapt==1.12.1
